finish lihaoyi hands on scala

+++ demo-js/js/src/main/scala/com/peknight/demo/js/lihaoyi/handson/crossbuilds/simple/Platform.scala
@@ -4,6 +4,6 @@ import scala.scalajs.js
 
 object Platform {
   def format(ts: Long) = {
-    new js.Date(ts).toISOString()
+    new js.Date(ts.toDouble).toISOString()
   }
 }
